백엔드 폴더명 변경
