--- backend/controllers/auth.js
@@ -1,56 +1,0 @@
-import jwt from 'jsonwebtoken';
-import bcrypt from 'bcrypt';
-import 'express-async-errors';
-import * as userRepository from '../data/auth.js';
-import { config } from '../config.js';
-
-export async function signup(req, res) {
-  const { email, password, username, name } = req.body;
-  const found = await userRepository.findByEmail(email);
-  
-  if (found) {
-    return res.status(409).json({ message: `${username} already exists` });
-  }
-
-  const hashed = await bcrypt.hash(password, config.bcrypt.saltRounds);
-  const userId = await userRepository.createUser({
-    email,
-    password: hashed,
-    username,
-    name,
-  });
-  const token = createJWTToken(userId);
-  res.status(201).json({ token, username });
-}
-
-export async function login(req, res) {
-  const { email, password } = req.body;
-  const user = await userRepository.findByEmail(email);
-  if (!user) {
-    return res.status(401).json({ message: 'Invalid user or password'});
-  }
-  const isValidPassword = await bcrypt.compare(password, user.password);
-  if (!isValidPassword) {
-    return res.status(401).json({ message: 'Invalid user or password '});
-  }
-  const token = createJWTToken(user.id);
-  res.status(200).json({ token, uername });
-}
-
-export async function me(req, res, next) {
-  const user = await userRepository.findById(req.userId);
-  if (!user) {
-    return res.status(404).json({ message: 'User not found' });
-  }
-  res.status(200).json({ token: req.token, email: user.email });
-}
-
-function createJWTToken(id){
-  return jwt.sign(
-    { id }, 
-    config.jwt.secretKey, 
-    {
-    expiresIn: config.jwt.expiresInSec,
-    }
-  );
-}
