+++ backend/middleware/validator.js
@@ -1,11 +1,9 @@
 import { validationResult } from 'express-validator';
 
 export const validate = (req, res, next) => {
-  const errors = validationResult(req);
-
-  if (errors.isEmpty()) {
-    return next();
-  }
-
-  return res.status(400).json({ message: errors.array()[0].msg });
+    const errors = validationResult(req);
+    if(errors.isEmpty()){
+        return next();
+    }
+    return res.status(400).json({ message: errors.array()[0].msg });
 }
