--- backend/server.js
@@ -1,29 +1,0 @@
-import express from 'express';
-import 'express-async-errors';
-import cors from 'cors';
-import morgan from 'morgan';
-import helmet from 'helmet';
-import feedsRouter from './routes/feeds.js';
-import authRouter from './routes/auth.js';
-import { config } from './config.js';
-
-const app = express();
-
-app.use(express.json());
-app.use(helmet());
-app.use(cors());
-app.use(morgan('tiny'));
-
-app.use('/feeds', feedsRouter);
-app.use('/auth', authRouter);
-
-app.use((req, res, next) => {
-  res.sendStatus(404);
-});
-
-app.use((error, req, res, next) => {
-  console.error(error);
-  res.sendStatus(500);
-});
-
-const server = app.listen(config.host.port);
